requirements.txt
anyio==3.6.2
attrs==22.2.0
autopep8==2.0.1
azure-core==1.26.2
azure-data-tables==12.4.1
azure-functions==1.12.0
azure-storage-blob==12.14.1
boltons==21.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
et-xmlfile==1.1.0
face==22.0.0
fastapi==0.89.1
glom==22.1.0
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
gspread==5.7.2
gspread-dataframe==3.3.0
h11==0.14.0
httplib2==0.21.0
httptools==0.5.0
idna==3.4
install==1.3.5
isodate==0.6.1
msrest==0.7.1
multidict==6.0.4
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.0
pandas==1.5.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
PyDrive==1.3.1
pygsheets==2.0.6
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==3.0.1
urllib3==1.26.13
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
